% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr.R
\name{easy_cmr}
\alias{easy_cmr}
\title{Simple cmr version with sorting-algorithm}
\usage{
easy_cmr(data)
}
\arguments{
\item{data}{currently the default data structure of pirst}
}
\value{
same data structure but with CMR values for X and Y
}
\description{
Simple cmr version with sorting-algorithm
}
