% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirst-package.R
\docType{package}
\name{pirst-package}
\alias{pirst}
\alias{pirst-package}
\title{pirst: Permuted-Isotonic-Regression for State Trace Analysis}
\description{
Non-parametric algorithm based on isotonic regresion to evaluate evidence of a state-trace plot for the hypotheses of a single dimension or two dimensions.
}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes.titz@gmail.com} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID})

Authors:
\itemize{
  \item Lukas Rommel
}

}
\keyword{internal}
