% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pirst_test}
\alias{pirst_test}
\title{PIRST significance test}
\usage{
pirst_test(data, nperms = 100, nboot = 100, conf = 0.95)
}
\arguments{
\item{data}{data set with columns:}

\item{nperms}{number of permutations}

\item{nboot}{number of bootstrap samples}

\item{conf}{confidence (1 - Alpha)}
}
\description{
PIRST significance test
}
